+++ appseed/ca/collection/collection_pointer_list.cpp
@@ -344,7 +344,7 @@ void pointer_list::dump(dump_context & dumpcontext) const
 {
    ::radix::object::dump(dumpcontext);
 
-   dumpcontext << "with " << LONGLONG(m_nCount) << " elements";
+   dumpcontext << "with " << m_nCount << " elements";
    if (dumpcontext.GetDepth() > 0)
    {
       POSITION pos = get_head_position();
