+++ appseed/ca/collection/collection_string_list.cpp
@@ -427,7 +427,7 @@ void string_list::dump(dump_context & dumpcontext) const
 {
    ::radix::object::dump(dumpcontext);
 
-   dumpcontext << "with " << LONGLONG(m_nCount) << " elements";
+   dumpcontext << "with " << m_nCount << " elements";
    if (dumpcontext.GetDepth() > 0)
    {
       POSITION pos = get_head_position();
